Allow ConvertFunctionsToVirtualList to be indexed with slices and longs (python2.x)
It may seem silly to allow longs, (after all, a PDF with that many pages has other problems),
but 1L should be valid as small longs should be equivalent to ints and some APIs may return
longs.

Slices are a little more obvious -- It allows the "VirtualList" objects to behave a little
more like regular lists.