Add abbreviated short-names for filters
After investigating an odd error:

# NotImplementedError: unsupported filter /Fl

I saw the PDF had this line:

# <</Filter/Fl/First 12/Length 828/N 2/Type/ObjStm>>stream

But most objects had:

# <</Size 306/Filter/FlateDecode/Length 947/Type/XRef/W[1 3 1]

After looking at the filters.py file, I saw the short-names were not added. After modifying the filters.py, the code is up and running again.

A thanks to Matthew Weiss for helping with this as well.