MAINT: Split generic.py (#1229)
The aim of this refactoring PR is to explicitly define the interface of `PyPDF2.generic` via `__all__` and to structure this big submodule more. I hope this makes it easier to test / expand in future if necessary. Smaller modules should have less merge conflicts.

This PR should not change anything for users of PyPDF2.