ENH: Rename resources deterministically in merge_page (#1543)
This fixes #1532 by adjusting the procedure used for renaming resources in `merge_page`, so that resources that have the same name (but different contents) are renamed in a deterministic/reproducible way.

When merging pages that both have a resource of the same name (say, `/Example`) but different values, the resource from the second page will now be renamed to `/Example-0`. Previously, it would be renamed with a random UUIDv4 (e.g. `/Exampledbbbe7cb-5f34-4061-b863-41919b326b49`) which would be different on every run, even if the inputs were identical.

The renamed name may already exist if a PDF is carefully/maliciously crafted, in which case `/Example-1`, `/Example-2`, etc. are tried until an appropriate name is found. If any of these options have the same value, that name/value is reused.

The `/Example-0` pattern is short and sweet, but can easily be changed, e.g. some options off the top of my head:

- `/Example-pypdf-merged-0` for explanation about where it comes from
- `/Example-9c4beb01-0` where the `9c4beb01` is some marker hardcoded into pypdf's code, to reduce the chance of a "normal" PDF requiring `idx > 0` (only truly maliciously crafted ones)