ENH: Add PdfWriter.remove_objects_from_page(page: PageObject, to_delete: ObjectDeletionFlag) (#1648)
This fixes remove_text to set contents as indirect_objects in accordance with the PDF specification.
It wipes out text in XObject forms as well.

The same issues were fixed for remove_images()

Finally, the new method
    PdfWriter.remove_objects_from_page(page: PageObject, to_delete: ObjectDeletionFlag)
was created. This allows a more fine-granular control of what to delete. It also is easy to expand via the to_delete flag.

Closes #1644
Closes #1650