Bug: add_named_destination() maintains named destination list sort order (#1930)
To quote the "Table 36 – Entries in a name tree node dictionary" the part about the `Names` key:

> The keys shall be sorted in lexical order, as described below. [...]
>
> The Names entries in the leaf (or root) nodes shall contain the tree’s keys and their
associated values, arranged in key-value pairs and shall be sorted lexically in ascending order by key. Shorter
keys shall appear before longer ones beginning with the same byte sequence. Any encoding of the keys may
be used as long as it is self-consistent; keys shall be compared for equality on a simple byte-by-byte basis.

Fixes #1927