MAINT: Update upload-artifact and download-artifact actions from v3 to v4 (#2352)
This is a general fix which covers #2344, #2345 and #2350, including some incompatible changes as detailed in #2344.

The key point is that both actions have to be updated at the same time to avoid conflicts. Additionally, artifacts of the same name are not supported anymore, thus I decided to go the recommended route from https://github.com/actions/upload-artifact?tab=readme-ov-file#not-uploading-to-the-same-artifact to include the matrix parameters inside the artifact names.